improvement: handle ubuntu installation methods
On Ubuntu with package based installation, we currently get nginx v1.4 with the option to install nginx-extras and use some advanced configuration options (some of which we need for hardening).

When the user chooses to install from source or official nginx repo, this is not possible. For both cases, we won't use advanced options which are often not available. Installation from source has to be handled separately in the future with better customization.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>